+++ .eslintrc.js
@@ -1,12 +1,19 @@
 module.exports = {
-  extends: ['@mate-academy/eslint-config-react-typescript', 'plugin:cypress/recommended'],
+  extends: ['@mate-academy/eslint-config-react', 'plugin:cypress/recommended'],
   rules: {
+    'import/no-extraneous-dependencies': ['error', {
+      devDependencies: true,
+      optionalDependencies: false,
+      peerDependencies: false,
+    }],
+    'react/prop-types': 0,
     'max-len': ['error', {
       ignoreTemplateLiterals: true,
       ignoreComments: true,
     }],
-    'jsx-a11y/label-has-associated-control': ["error", {
-      assert: "either",
+    'jsx-a11y/label-has-associated-control': ['error', {
+      assert: 'either',
     }],
+    'jsx-a11y/control-has-associated-label': 'off',
   },
 };
