+++ README.md
@@ -1,4 +1,4 @@
-# React Tabs
+# React Tabs (JS)
 
 Implement the `App` and `Tabs` components to render and switch between given `tabs`.
 
@@ -21,6 +21,5 @@ Implement the `App` and `Tabs` components to render and switch between given `ta
 ## Instructions
 
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline).
-- Use the [React TypeScript cheat sheet](https://mate-academy.github.io/fe-program/js/extra/react-typescript).
 - Open one more terminal and run tests with `npm test` to ensure your solution is correct.
 - Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_tabs/) and add it to the PR description.
