+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "react_tabs",
+  "name": "react_tabs-js",
   "homepage": ".",
   "version": "0.1.0",
   "keywords": [],
@@ -17,7 +17,7 @@
   "devDependencies": {
     "@cypress/webpack-dev-server": "^1.8.4",
     "@mate-academy/cypress-tools": "^1.0.5",
-    "@mate-academy/eslint-config-react": "*",
+    "@mate-academy/eslint-config-react": "^0.0.11",
     "@mate-academy/eslint-config-react-typescript": "*",
     "@mate-academy/scripts": "^1.2.8",
     "@mate-academy/students-ts-config": "*",
@@ -61,8 +61,8 @@
     ]
   },
   "mateAcademy": {
-    "_comment": "Replace 'reactTypescript' with 'react' if you want use React without Typescript",
-    "projectType": "reactTypescript",
+    "_comment": "Replace 'react' with 'reactTypescript' if you want use React with Typescript",
+    "projectType": "react",
     "tests": {
       "_comment": "Add `cypressComponents: true` to enable component tests",
       "cypressComponents": true,
