+++ src/components/Tabs/Tabs.spec.jsx
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-unused-expressions, react/jsx-filename-extension */
+/* eslint-disable react/jsx-filename-extension */
 import { mount } from '@cypress/react';
 import { Tabs } from './Tabs';
